[power] Move suspend()/resume() methods.
Instead of putting them in struct device_driver (which few, if any drivers
use directly), put them in the controlling bus_type of the device (which
are currently responsible for claiming the methods and forwarding the calls
to the bus-specific driver anyway).

This will save 8 bytes per driver instance, which isn't that much, but it's
something. It also makes it more obvious to the reader what is going on. 
And, it makes for easier bus-level defaults in the case the device has no
driver attached. 

The old calls remain until all instances have been fixed up.