[kobject] Support unlimited name lengths.
Add ->k_name pointer which points to the name for a kobject. By default, this
points to ->name (the static name array). 
 
Users of kobjects may use the helper function kobject_set_name() (and are 
encouraged to do so in all cases). This function will determined whether or
not the name is short enough to fit in ->name. If so, great. 

Otherwise, a dyanamic string is allocated and the name is stored there. 
->k_name will point to that, and will be freed when the kobject is released. 

kobject_set_name() may take a format string, like:

	kobject_set_name(kobj,"%s%d",base_name,id); 

and will behave as expected (will put in ->name, unless it's too long, in 
which case a new string will be allocated and it will be stored in there).