[PATCH] Fixes to allow ARM to build in the standard tree
In addition to the MODULE_ALIAS_LDISC patch, the following changes to
generic code are needed to allow ARM to build in Linus' tree.

The first is to add PT_SINGLESTEP to ptrace.h so we do the right thing
when adding/removing breakpoint instructions into processes (see
arch/arm/kernel/{signal.c,ptrace.c} for usage.)

The second is needed because pmd_clear() needs to flush the pmd.
However, we can't include tlbflush.h into pgtable.h without causing
a circular dependency (tlbflush.h needs vm_area_struct and mm_struct
which are in mm.h, which needs pgtable.h.)  swapfile.c seems to be the
only file affected.