[PATCH] dac960 devfs_name initialisation fix
From: Andrey Borzenkov <arvidjaar@mail.ru>

DAC960.c does it incorrectly (at least in 2.6.0-test4). It will create
_directory_ /dev/rd/cNdM making it impossible to create compat block device
entry with the same name.

The right thing it to create separate directory for each controller/target as 
in attached trivial patch (untested due to lack of hardware). You will need
devfsd support for this but then you will need it for cciss or cpqarray as 
well and possibly for others. Which returns us to the problem of devfsd 
maintenance ...