[PATCH] CPU scheduler balancing fix
From: Nick Piggin <piggin@cyberone.com.au>

The patch changes the imbalance required before a balance to 25% from 50% -
as the comments intend.  It also changes a case where the balancing
wouldn't be done if the imbalance was >= 25% but only 1 task difference.

The downside of the second change is that one task may bounce from one cpu
to another for some loads.  This will only bounce once every 200ms, so it
shouldn't be a big problem.

(Benchmarking results are basically a wash - SDET is increased maybe 0.5%)